refactor: toast 관련 파일 관리 방식 변경
