+++ packages/user/src/App.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from 'react-router-dom';
-import { Toaster } from 'src/components/ui/toaster';
+import ToasterStack from 'src/components/common/toast/ToasterStack';
 import AppProviders from 'src/libs';
 import router from 'src/routes/router';
 
@@ -8,7 +8,7 @@ export default function App() {
 		<AppProviders>
 			{/* TODO: create global loading component */}
 			<RouterProvider router={router} fallbackElement={<>loading...</>} />
-			<Toaster />
+			<ToasterStack />
 		</AppProviders>
 	);
 }
