+++ packages/user/src/components/common/toast/Toast.tsx
@@ -1,6 +1,6 @@
 import * as ToastPrimitives from '@radix-ui/react-toast';
 import { cva, type VariantProps } from 'class-variance-authority';
-import * as React from 'react';
+import React from 'react';
 import cn from 'src/utils/cn';
 
 const ToastProvider = ToastPrimitives.Provider;
